requirements.txt
pyhyphen
pronouncing
jellyfish
